move method to utils, used in multiple tests
