+++ openmdao/utils/general_utils.py
@@ -8,7 +8,7 @@ import unittest
 from mock import Mock
 from fnmatch import fnmatchcase
 from six import string_types
-from six.moves import range
+from six.moves import range, cStringIO as StringIO
 from collections import Iterable
 import numbers
 
@@ -349,3 +349,29 @@ def find_matches(pattern, var_list):
     elif pattern in var_list:
         return [pattern]
     return [name for name in var_list if fnmatchcase(name, pattern)]
+
+
+def run_model(prob):
+    """
+    Call `run_model` on problem and capture output.
+
+    Parameters
+    ----------
+    prob : Problem
+        an instance of Problem
+
+    Returns
+    -------
+    string
+        output from calling `run_model` on the Problem, captured from stdout
+    """
+    stdout = sys.stdout
+    strout = StringIO()
+
+    sys.stdout = strout
+    try:
+        prob.run_model()
+    finally:
+        sys.stdout = stdout
+
+    return strout.getvalue()
